requirements.txt
# Core Google Cloud libraries
google-cloud-documentai>=2.21.0
google-cloud-storage>=2.10.0
google-api-core>=2.11.1

# PDF processing
PyMuPDF>=1.23.0
fitz>=0.0.1.dev2

# Image processing  
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Optional: For advanced image preprocessing
scikit-image>=0.21.0
imageio>=2.31.0

streamlit==1.28.0
google-auth==2.23.3
nameparser==1.1.2
python-dotenv==1.0.0
openpyxl==3.1.2

# Database dependencies
psycopg2-binary==2.9.7
sqlalchemy==2.0.23